FirstPush-Diplom

+++ ewm-service/src/main/java/ru/practicum/service/categories/CategoriesServiceImpl.java
@@ -1,17 +1,22 @@
 package ru.practicum.service.categories;
 
 import lombok.AllArgsConstructor;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import ru.practicum.client.EwmClient;
 import ru.practicum.dto.CategoryDto;
 import ru.practicum.dto.NewCategoryDto;
 import ru.practicum.exception.AlreadyExistsException;
 import ru.practicum.exception.NotFoundException;
 import ru.practicum.model.Category;
+import ru.practicum.model.Event;
 import ru.practicum.repository.CategoryRepository;
+import ru.practicum.repository.EventRepository;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -20,31 +25,37 @@ import java.util.stream.Collectors;
 public class CategoriesServiceImpl implements CategoriesService {
 
     private final CategoryRepository categoryRepository;
+    private final EventRepository eventRepository;
+    private final EwmClient ewmClient;
 
     @Transactional
     @Override
     public CategoryDto create(NewCategoryDto dto) {
-        return CategoryMapper.toCategoryDto(categoryRepository.save(CategoryMapper.toCategory(dto)));
+        if (categoryRepository.findByName(dto.getName()).isPresent()) {
+            throw new AlreadyExistsException("Категория с таким названием уже существует");
+        }
+        return CategoryMapper.toCategoryDto(categoryRepository.save(CategoryMapper.toCategoryfromNew(dto)));
     }
 
     @Transactional
     @Override
     public void delete(Long id) {
+        List<Event> events = eventRepository.findByCategoryId(id);
+        if (events.size() > 0) {
+            throw new DataIntegrityViolationException("Категория связана с событиями");
+        }
         categoryRepository.deleteById(id);
     }
 
     @Transactional
     @Override
-    public CategoryDto update(CategoryDto dto) {
+    public CategoryDto update(Long catId, CategoryDto dto) {
+        Category category = CategoryMapper.toCategoryUpdate(dto, catId);
         Category cat = categoryRepository
-                .findById(dto.getId())
+                .findById(catId)
                 .orElseThrow(() -> new NotFoundException(
                         String.format("Категория с id=%d не найдена", dto.getId())));
-        if (categoryRepository.findByName(dto.getName()).isPresent()) {
-            throw new AlreadyExistsException(
-                    String.format("Категория - %s уже присутствует", dto.getName()));
-        }
-        cat.setName(dto.getName());
+        cat.setName(category.getName());
         return CategoryMapper.toCategoryDto(categoryRepository.save(cat));
     }
 
@@ -56,11 +67,12 @@ public class CategoriesServiceImpl implements CategoriesService {
     }
 
     @Override
-    public CategoryDto getById(long catId) {
+    public CategoryDto getById(long catId, HttpServletRequest httpServletRequest) {
         Category cat = categoryRepository
                 .findById(catId)
                 .orElseThrow(() -> new NotFoundException(
                         String.format("Категория с id=%d не найдена", catId)));
+        ewmClient.addHit(httpServletRequest);
         return CategoryMapper.toCategoryDto(cat);
     }
 }
